+++ src/main/java/roomescape/domain/reservation/service/ReservationService.java
@@ -1,0 +1,39 @@
+package roomescape.domain.reservation.service;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import roomescape.domain.reservation.domain.Reservation;
+import roomescape.domain.reservation.domain.repository.ReservationRepository;
+
+import java.util.List;
+
+@Service
+public class ReservationService {
+
+    private final ReservationRepository reservationRepository;
+
+    public ReservationService(ReservationRepository reservationRepository) {
+        this.reservationRepository = reservationRepository;
+    }
+
+    @Transactional
+    public Reservation save(Reservation reservation) {
+        long id = reservationRepository.save(reservation);
+        return findById(id);
+    }
+
+    @Transactional(readOnly = true)
+    public Reservation findById(long id) {
+        return reservationRepository.findById(id);
+    }
+
+    @Transactional(readOnly = true)
+    public List<Reservation> findAll() {
+        return reservationRepository.findAll();
+    }
+
+    @Transactional
+    public void delete(Long id) {
+        reservationRepository.delete(id);
+    }
+}
