+++ src/main/java/roomescape/domain/time/api/dto/TimeResponse.java
@@ -1,0 +1,20 @@
+package roomescape.domain.time.api.dto;
+
+public class TimeResponse {
+
+    private final Long id;
+    private final String startAt;
+
+    public TimeResponse(Long id, String startAt) {
+        this.id = id;
+        this.startAt = startAt;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getStartAt() {
+        return startAt;
+    }
+}
