+++ src/main/java/roomescape/presentation/ui/admin/AdminPageController.java
@@ -1,0 +1,28 @@
+package roomescape.presentation.ui.admin;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class AdminPageController {
+
+    @GetMapping("/")
+    String adminPage() {
+        return "admin/index";
+    }
+
+    @GetMapping("/admin/reservation")
+    public String adminReservationPage() {
+        return "admin/reservation";
+    }
+
+    @GetMapping("/admin/theme")
+    public String adminThemePage() {
+        return "admin/theme";
+    }
+
+    @GetMapping("/admin/time")
+    public String adminTimePage() {
+        return "/admin/time";
+    }
+}
