--- src/main/java/roomescape/theme/service/ThemeService.java
@@ -1,38 +1,0 @@
-package roomescape.theme.service;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import roomescape.theme.domain.Theme;
-import roomescape.theme.domain.ThemeRepository;
-
-import java.util.List;
-
-@Service
-public class ThemeService {
-
-    private final ThemeRepository themeRepository;
-
-    public ThemeService(ThemeRepository themeRepository) {
-        this.themeRepository = themeRepository;
-    }
-
-    @Transactional
-    public Theme save(Theme theme) {
-        Long id = themeRepository.save(theme);
-        return findById(id);
-    }
-
-    @Transactional(readOnly = true)
-    public Theme findById(Long id) {
-        return themeRepository.findById(id);
-    }
-
-    @Transactional(readOnly = true)
-    public List<Theme> findAll() {
-        return themeRepository.findAll();
-    }
-
-    public void delete(Long id) {
-        themeRepository.delete(id);
-    }
-}
