--- src/main/java/roomescape/time/error/exception/TimeException.java
@@ -1,15 +1,0 @@
-package roomescape.time.error.exception;
-
-import roomescape.global.error.exception.UserException;
-
-public class TimeException extends UserException {
-
-    public TimeException(ErrorCode errorCode) {
-        super(errorCode.getStatus(), errorCode.getErrorMessage());
-    }
-
-    @Override
-    public int getStatus() {
-        return super.getStatus();
-    }
-}
