+++ src/test/java/roomescape/ThemeTest.java
@@ -10,10 +10,10 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.theme.domain.Theme;
-import roomescape.theme.error.exception.ErrorCode;
-import roomescape.theme.error.exception.ThemeException;
-import roomescape.theme.presentation.dto.ThemeRequest;
+import roomescape.domain.theme.domain.Theme;
+import roomescape.domain.theme.error.exception.ErrorCode;
+import roomescape.domain.theme.error.exception.ThemeException;
+import roomescape.domain.theme.api.dto.ThemeRequest;
 
 import java.util.HashMap;
 import java.util.Map;
